pyproject.toml
[tool.poetry]
name = "uaip"
version = "0.1.0"
description = ""
authors = ["Yung-Jin (Joey) Hu <yungjinhu@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
attrs = "^19.3.0"
twisted = "^20.3.0"
janus = "^0.5.0"
aiohttp = "^3.6.2"
pyzmq = "^19.0.2"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
pylint = "^2.5.3"
ipython = "^7.17.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
